Update README.md
Bitcoin without the mathematics.
A token used as a utility type raising capital for 
AutoGrowBoxes and other Technology Used in conjuction
to form the foundation. why I used 25% x 4 = 100% a budget
for the mars preserve research foundation being able to work in the workspace while getting more money for github services.